moving the institution class to the package models
